fix(tiler-sharp): when resizing preserve input data type (#3417)
### Motivation

When applying a crop on a 1-band `unit8` the crop should not change the
datatype to `float32`.

Changing the data type to `float32` changes the default `color-ramp` to
one used by elevation, which causes hillshades to be rendered green

Igor hillshade before Fix:


![13_7998_5185-color-ramp_before](https://github.com/user-attachments/assets/857caa96-5afb-4f0a-843c-3b70666f0063)

After Fix:

![13_7998_5185-color-ramp_after](https://github.com/user-attachments/assets/01f89a0f-1f5d-41c8-bc1a-0d0e9439f654)



### Modifications

Ensure the input data type is preserved when output datatype is made

### Verification

Unit tests and raw tiles rendered.