perf(tiler-sharp): split greyscale  color ramps into specalized function (#3418)
### Motivation

Currently when expanding a `uint8` color into a 4-band RGBA
`uint8,uint8,uint8,uint8` the `color-ramp` is used which is not designed
for a direct color mapping.

For a 1024x1024 tile the current color ramp implementation takes approx
7ms to convert a 1-band to 4 band grey scale, using a direct band
expansion drops the time taken to approx 1ms.

This pull requests creates a simple color mapper where given a input
value (eg 240) it is propagated to the R,G and B bands of the output eg
`240,240,240` creating a grey scale image from a one band input.


### Modifications

Create a specialised `Colorizer` for one band inputs that is
significantly (5x) faster than using a color ramp directly

### Verification

Unit tests to validate inputs and outputs are the same between
implementations