fix(cli): update landing deploy script with new cli-config package BM-1260 (#3431)
### Motivation

In [another piece of work], we re-located commands from the `cli`
package to a new `cli-config` package. In doing so, we unknowingly broke
the `deploy.mjs` script that lives in the `landing` package.

[another piece of work]: https://github.com/linz/basemaps/pull/3428

### Modifications

This work corrects the dependency issue causing the `deploy.mjs` script
to break. This work also adjusts the `landing` package's configuration
so that its `scripts` directory is included in error-checking and
type-checking.

- **packages/landing**

  - **package.json**

Added missing dependencies required by `scripts/deploy.mjs` to the
`devDependencies` property:

    -  `@basemaps/cli-config`
    -  `@types/mime-types`

  - **tsconfig.json**

1. Updated the `include` property to capture source files within the
`scripts` directory during compilation.
3. Updated the `rootDir` property to the package's root (`.`) as it is
the closest common directory for both the `scripts` and `src`
directories.
4. Enabled the `checkJs` property to enable error-checking for non-ts
files (i.e. `scripts/deploy.mjs`).
5. Extended the `references` property array to capture the `landing`
package's [dependence] on the `cli-config` package.

  - **scripts/deploy.mjs**

With error-checking now enabled for this file, the source code threw
some errors. I have modified the logic to resolve the errors. I
anticipate that the modifications will not have augmented the logic, but
I will have to verify this.

### Verification

check landing deploy script working

[dependence]:
https://github.com/linz/basemaps/blob/e527a04fec65c82f8577642493e8fcb475762243/packages/landing/scripts/deploy.mjs#L1

---------

Co-authored-by: Tawera Manaena <TManaena@linz.govt.nz>