fix: ensure all linzjs packages are correctly labeled as deps (#3439)
### Motivation

There is a weird behaviour in this monorepo that when importing packages
that are present inside the monorepo can be imported freely anywhere
inside the monorepo, but when the package is published it needs to exist
inside the package.json or it will not be installed correctly.

We have a script that detects when this happens
`scripts/detect.unlinked.dep.mjs` but it was only functioning for
`@basemaps/*` packages in this repository there are a few `@linzjs/`
packages.

### Modifications

Ensure linzjs packages are also checked as part of CI
Add the missing links between a few packages and `@linzjs/geojson` and
`@linzjs/metrics`

### Verification

`detect.unlinked.dep.mjs` found these packages once it was updated to
include looking for `@linzjs`

Fixes #3438