fix(infra): increase lambda tiler memory to 3072MB (#3459)
### Motivation

We are experiencing an elevated error rate from **out of memory**
exceptions. We believe this might be related to a large collection of
new imagery we are publishing.

More details avaliable via Slack [here].

### Modifications

As a temporary measure, we have increased the memory capacity of the
main lambda to `3072MB`. We will look to revert this in the near future
as this will likely have increased cost implications.

[here]: https://linz.slack.com/archives/CNKCWCD3M/p1750036843048369