fix(cli-vector): revert polylabel version to 1.1.0 BM-1309 (#3467)
### Motivation

The Airport labels for our `topographic-v2` basemap are in the wrong
place. This is because we use a newer version of
[polylabel](https://github.com/mapbox/polylabel) in our `cli-vector`
package (2.0.1), compared to `basemaps-team` (1.1.0). Evidently, with
the default precision value of `1.0`, the algorithm renders different
results between the two versions.

| [topographic] | [topographic-v2] |
| - | - |
| ![][img_v1] | ![][img_v2] |

[img_v1]:
https://github.com/user-attachments/assets/e818542f-345f-4bb2-beec-d23feb6ce0cd
[img_v2]:
https://github.com/user-attachments/assets/4cb2b91e-79a1-4a09-9f39-97990a9504b8

[topographic]:
https://basemaps.linz.govt.nz/@-41.3270061,174.8124381,z14.48?style=topographic&i=topographic
[topographic-v2]:
https://basemaps.linz.govt.nz/@-41.3270061,174.8124381,z14.48?style=topographic-v2&i=topographic-v2

Adjusting the precision value (e.g. `0.000001`) with polylabel version
`2.0.1` does render precise pole of inaccessiblity points. However, the
locations produced with polylabel version `1.1.0` are more ideal for
where we want our labels to appear.

More discussion in this [Slack
thread](https://linz.slack.com/archives/C5TCTQUJG/p1750735625977969).

### Modifications

#### Packages

- **cli-vector**

   - Revert the `polylabel` dependency's version from `2.0.1` to `1.1.0`
- Revert the `@types/polylabel` dev dependency's version from `1.1.3` to
`1.0.5`

### Verification

I've captured the following screenshots in QGIS showing the Airport
label point for our `topographic` and `topographic-v2` basemaps. The
last column shows where the new Airport label point will be. The same
location as it is in our `topographic` basemap.

| [topographic] (Ideal) | [topographic-v2] (Wrong) | topographic-v2
(Corrected) |
| - | - | - |
| ![][img_qgis_v1] | ![][img_qgis_v2] | ![][img_qgis_updated] |

[img_qgis_v1]:
https://github.com/user-attachments/assets/d5f233f1-9694-4ec6-b852-448728ebdd68
[img_qgis_v2]:
https://github.com/user-attachments/assets/703d88a9-e6fe-40ca-81ec-561c92559b67
[img_qgis_updated]:
https://github.com/user-attachments/assets/72bd11a0-dcce-44f8-964d-d9a7ba6cba0c