fix: intermittent unit tests (#3503)
### Motivation

A unit test was comparing two ULID's time components which could be
generated on the same millisecond, which caused the test to fail as it
expects the first id to be older than the second.

### Modifications

make the first id much older than the current time.

### Verification

<!-- TODO: Say how you tested your changes. -->

ran the test 100x to see if it would fail again, was failing approx 1 in
5 before the change.