fix(config): DynamoDB error caused by 'undefined' value in config tilesets BM-1373 (#3528)
### Motivation

Our basemaps-config repo's [Deploy Nonprod] job is failing.

[Deploy Nonprod]:
https://github.com/linz/basemaps-config/actions/runs/17960304459/job/51082025343#step:4:177

The failure is due to our code assigning `{ outputs: undefined }`
key-value pairs to `ConfigTileset` objects. When our code attempts to
upload such objects, DynamoDB throws an error as it does not support
object properties with `undefined` values.

### Modifications

- **packages/config**

  - **src/memory**

    - `memory.config.ts`

Adjusted our code so that it only assigns the value returned from the
`addDefaultOutputPipelines` function if the value is not `null` or
`undefined`.

### Verification

- **packages/config**

  - `memory.config.test.ts`

Added a test case to the existing suite to ensure we no longer assign `{
outputs: undefined }` to `ConfigTileset` objects by default.