fix(landing): keep gridless default (#3536)
This PR assists [this PR].

[this PR]: https://github.com/linz/basemaps-config/pull/1327

---

### Motivation

In a recent [piece of work], we switched our `topo-raster` tileset to
serve the 600 DPI `gridded` map sheets. This changed the imagery for our
customers who were consuming our original `topo-raster` endpoint. We've
decided we want to switch the original endpoint back to serving the
`gridless` map sheets.

[piece of work]: https://github.com/linz/basemaps-config/pull/1319

### Modifications

<!-- TODO: Say what changes you made. -->

<!-- TODO: Attach screenshots if you changed the UI. -->

### Verification

<!-- TODO: Say how you tested your changes. -->