feat(landing): display downloadable mbtiles message for vector basemaps BM-1330 (#3478)
### Motivation

We recently added a new `export` API Endpoint. This endpoint allows
users to download the production-side `MBTiles` for any of our vector
basemaps. Although useful, we don't want to openly advertise the
endpoint as each request incurs a cost. Instead, we have opted to
display a call-to-action message for users to _contact us_ first.

### Modifications

- **packages**

   - **landing**
   
- Added the following message to the right-hand sidebar that appears
only when the selected layer is a vector basemap:
        
        > **MBTiles**
> Request downloadable MBTiles for use in offline or mobile apps using
the **Contact us** button below.

### Verification

The following screenshots show what content our application renders
based on the selected layer. As shown, the new message appears only for
vector basemaps.

| Aerial | Topographic | Topolite |
| - | - | - |
| ![][aerial] | ![][topographic] | ![][topolite] |
| Raster: _Message hidden_ | Vector: _Message displayed_ | Vector:
_Message displayed_ |

[aerial]:
https://github.com/user-attachments/assets/5ef66e65-8135-4ba6-bb51-433a24eaf03b
[topographic]:
https://github.com/user-attachments/assets/05873b30-fad1-4e62-a0fe-b07e16d0f9f8
[topolite]:
https://github.com/user-attachments/assets/01df4e14-fa37-4ff4-85bb-80c7b05a8840