feat: support default outputs pipelines for backwards compatibility (#3561)
### Motivation

Output pipelines allow admins to customise how datasets are served to
the users, common pipelines are `color-ramp` to convert elevation data
into RGBA or `terrain-rgb` to convert elevation to terrain-rgb

If there is only one output pipeline, then it will be picked as the
default option for the tile, so all endpoints will assume the use of the
pipeline.

If a second pipeline is added, the first link will now be broken as
there is no default for the system to choose

### Modifications

Adds a `default: boolean` option to the pipeline outputs to allow admins
to configure what the default pipeline is

### Verification

unit tests.

---------

Co-authored-by: Tawera Manaena <TManaena@linz.govt.nz>