fix: use the configuration version not time to reset cache (#3553)
### Motivation

Loading a huge amount of tiffs to compile a configuration takes a long
time in some cases 30+ minutes, when recompiling these configs as part
of pull requests parts of the config can be cached so every individual
tiff file does not need to be loaded.

We current store that cache for a month and reset it on the first of the
month, there is nothing stopping us storing that cache for longer as the
first pull requests of the month take a long time to process as they are
starting from a cold cache.

### Modifications

Move the cache to only reset if a configuration change has been
detected.

### Verification

Unit tests/run locally